move njit function to operation.py and add unittest for each tranform function
